+++ src/backend/vulkan/BufferVk.h
@@ -40,7 +40,7 @@ namespace backend { namespace vulkan {
                            nxt::BufferUsageBit targetUsage) const;
 
       private:
-        void SetSubDataImpl(uint32_t start, uint32_t count, const uint32_t* data) override;
+        void SetSubDataImpl(uint32_t start, uint32_t count, const uint8_t* data) override;
         void MapReadAsyncImpl(uint32_t serial, uint32_t start, uint32_t count) override;
         void MapWriteAsyncImpl(uint32_t serial, uint32_t start, uint32_t count) override;
         void UnmapImpl() override;
